re-write tests as individual files
